Removes warnings.

+++ src/viewport.cpp
@@ -11,8 +11,8 @@ void Badger::Viewport::draw(Renderer* renderer) {
   renderer->bindTexture(map->spriteSheet()->texture());
 
   // Render all background tiles
-  for (int y = map->height() - 1; y >= 0; y--) {
-    for (int x = 0; x < map->width(); x++) {
+  for (long y = map->height() - 1; y >= 0; y--) {
+    for (long x = 0; x < (long)map->width(); x++) {
       Tile* tile = map->tile(x, y);
       if (!tile->passable) {
         continue;
@@ -29,8 +29,8 @@ void Badger::Viewport::draw(Renderer* renderer) {
   }
 
   // Render all foreground Tiles
-  for (int y = map->height() - 1; y >= 0; y--) {
-    for (int x = 0; x < map->width(); x++) {
+  for (long y = map->height() - 1; y >= 0; y--) {
+    for (long x = 0; x < (long)map->width(); x++) {
       Tile* tile = map->tile(x, y);
       if (tile->passable) {
         continue;
@@ -39,15 +39,15 @@ void Badger::Viewport::draw(Renderer* renderer) {
       map->spriteSheet()->textureCoordinates(tile->spriteIndex, coords);
 
       Sprite* sprite = map->spriteSheet()->sprite(tile->spriteIndex);
-      renderer->drawSquare(x*32.0f, y*32.0f + ((sprite->height - 32.0) / 2.0),
+      renderer->drawSquare(x*32.0f, y*32.0f + (((float)sprite->height - 32.0f) / 2.0f),
                            (float)sprite->width, (float)sprite->height,
                            coords[0], coords[1], coords[2], coords[3],
-                           -y*32.0f);
+                           -(double)y*32.0);
     }
   }
 
   // draw the actors
-  for (int x = 0; x < _world->actorCount(); x++) {
+  for (long x = 0; x < (long)_world->actorCount(); x++) {
     Badger::Actor* actor = _world->actor(x);
     renderer->bindTexture(actor->spriteSheet()->texture());
